107671: Split Theme model & ThemeConfig classes in separate files to prevent circular dependencies
