+++ src/app/shared/theme-support/theme.model.ts
@@ -1,57 +1,13 @@
 /* eslint-disable max-classes-per-file */
-import { Config } from './config.interface';
-import { hasValue, hasNoValue, isNotEmpty } from '../app/shared/empty.util';
-import { DSpaceObject } from '../app/core/shared/dspace-object.model';
-import { getDSORoute } from '../app/app-routing-paths';
-import { HandleObject } from '../app/core/shared/handle-object.model';
+import { hasValue, hasNoValue, isNotEmpty } from '../empty.util';
+import { DSpaceObject } from '../../core/shared/dspace-object.model';
+import { getDSORoute } from '../../app-routing-paths';
+import { HandleObject } from '../../core/shared/handle-object.model';
 import { Injector } from '@angular/core';
-import { HandleService } from '../app/shared/handle.service';
+import { HandleService } from '../handle.service';
 import { combineLatest, Observable, of as observableOf } from 'rxjs';
 import { map, take } from 'rxjs/operators';
-
-export interface NamedThemeConfig extends Config {
-  name: string;
-
-  /**
-   * Specify another theme to build upon: whenever a themed component is not found in the current theme,
-   * its ancestor theme(s) will be checked recursively before falling back to the default theme.
-   */
-  extends?: string;
-
-  /**
-   * A list of HTML tags that should be added to the HEAD section of the document, whenever this theme is active.
-   */
-  headTags?: HeadTagConfig[];
-}
-
-/**
- * Interface that represents a single theme-specific HTML tag in the HEAD section of the page.
- */
-export interface HeadTagConfig extends Config {
-  /**
-   * The name of the HTML tag
-   */
-  tagName: string;
-
-  /**
-   * The attributes on the HTML tag
-   */
-  attributes?: {
-    [key: string]: string;
-  };
-}
-
-export interface RegExThemeConfig extends NamedThemeConfig {
-  regex: string;
-}
-
-export interface HandleThemeConfig extends NamedThemeConfig {
-  handle: string;
-}
-
-export interface UUIDThemeConfig extends NamedThemeConfig {
-  uuid: string;
-}
+import { HandleThemeConfig, NamedThemeConfig, RegExThemeConfig, UUIDThemeConfig, ThemeConfig } from '../../../config/theme.config';
 
 export class Theme {
   constructor(public config: NamedThemeConfig) {
@@ -71,7 +27,7 @@ export class RegExTheme extends Theme {
   }
 
   matches(url: string, dso: DSpaceObject): Observable<boolean> {
-    let match;
+    let match: RegExpMatchArray;
     const route = getDSORoute(dso);
 
     if (isNotEmpty(route)) {
@@ -92,7 +48,7 @@ export class HandleTheme extends Theme {
 
   constructor(public config: HandleThemeConfig,
               protected handleService: HandleService
-              ) {
+  ) {
     super(config);
     this.normalizedHandle$ = this.handleService.normalizeHandle(this.config.handle).pipe(
       take(1),
@@ -133,9 +89,3 @@ export const themeFactory = (config: ThemeConfig, injector: Injector): Theme => 
     return new Theme(config as NamedThemeConfig);
   }
 };
-
-export type ThemeConfig
-  = NamedThemeConfig
-  | RegExThemeConfig
-  | HandleThemeConfig
-  | UUIDThemeConfig;
