+++ src/app/shared/theme-support/theme.service.ts
@@ -8,7 +8,8 @@ import { hasNoValue, hasValue, isNotEmpty } from '../empty.util';
 import { RemoteData } from '../../core/data/remote-data';
 import { DSpaceObject } from '../../core/shared/dspace-object.model';
 import { getFirstCompletedRemoteData, getFirstSucceededRemoteData, getRemoteDataPayload } from '../../core/shared/operators';
-import { HeadTagConfig, Theme, ThemeConfig, themeFactory } from '../../../config/theme.model';
+import { Theme, themeFactory } from './theme.model';
+import { ThemeConfig, HeadTagConfig } from '../../../config/theme.config';
 import { NO_OP_ACTION_TYPE, NoOpAction } from '../ngrx/no-op.action';
 import { followLink } from '../utils/follow-link-config.model';
 import { LinkService } from '../../core/cache/builders/link.service';
